in Klasse FragmentOne wird jetzt l und nicht mehr i übergeben und es wurde auf long angepasst

+++ app/src/main/java/com/fabian/android/aufgabenliste/Aufgaben.java
@@ -18,7 +18,7 @@ public class Aufgaben extends AppCompatActivity
         setContentView(R.layout.activity_aufgaben);
 
         Intent intent = getIntent();
-        int i = intent.getExtras().getInt("id");
+        long i = intent.getExtras().getLong("id");
 
         toolbar = (Toolbar) findViewById(R.id.toolbarVersion);
         setSupportActionBar (toolbar);
